requirements.txt
adafruit-io==2.8.0
bcrypt==4.3.0
bidict==0.23.1
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
dnspython==2.7.0
dotenv==0.9.9
Flask==3.1.0
flask-cors==5.0.1
Flask-SocketIO==5.5.1
h11==0.14.0
idna==3.10
importlib_metadata==8.6.1
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
paho-mqtt==2.1.0
PyJWT==2.10.1
pymongo==4.11.2
python-dotenv==1.0.1
python-engineio==4.12.0
python-socketio==5.13.0
requests==2.32.3
simple-websocket==1.1.0
urllib3==2.3.0
Werkzeug==3.1.3
wsproto==1.2.0
zipp==3.21.0
